+++ packages/autorest.gotest/test/integrationtest/output/compute/armcompute/galleryapplicationversions_client_example_test.go
@@ -35,7 +35,7 @@ func ExampleGalleryApplicationVersionsClient_BeginCreateOrUpdate() {
 		Location: to.Ptr("West US"),
 		Properties: &armcompute.GalleryApplicationVersionProperties{
 			PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-				EndOfLifeDate:      to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00Z"); return t }()),
+				EndOfLifeDate:      to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t }()),
 				ReplicaCount:       to.Ptr[int32](1),
 				StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 				TargetRegions: []*armcompute.TargetRegion{
@@ -72,9 +72,9 @@ func ExampleGalleryApplicationVersionsClient_BeginCreateOrUpdate() {
 	// 	Properties: &armcompute.GalleryApplicationVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryApplicationVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00+00:00"); return t}()),
+	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t}()),
 	// 			ExcludeFromLatest: to.Ptr(false),
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.5972568+00:00"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.597Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -110,7 +110,7 @@ func ExampleGalleryApplicationVersionsClient_BeginUpdate() {
 	poller, err := clientFactory.NewGalleryApplicationVersionsClient().BeginUpdate(ctx, "myResourceGroup", "myGalleryName", "myGalleryApplicationName", "1.0.0", armcompute.GalleryApplicationVersionUpdate{
 		Properties: &armcompute.GalleryApplicationVersionProperties{
 			PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-				EndOfLifeDate:      to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00Z"); return t }()),
+				EndOfLifeDate:      to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t }()),
 				ReplicaCount:       to.Ptr[int32](1),
 				StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 				TargetRegions: []*armcompute.TargetRegion{
@@ -147,9 +147,9 @@ func ExampleGalleryApplicationVersionsClient_BeginUpdate() {
 	// 	Properties: &armcompute.GalleryApplicationVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryApplicationVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00+00:00"); return t}()),
+	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t}()),
 	// 			ExcludeFromLatest: to.Ptr(false),
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.5972568+00:00"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.597Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -195,9 +195,9 @@ func ExampleGalleryApplicationVersionsClient_Get_getAGalleryApplicationVersionWi
 	// 	Properties: &armcompute.GalleryApplicationVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryApplicationVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00+00:00"); return t}()),
+	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t}()),
 	// 			ExcludeFromLatest: to.Ptr(false),
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.5972568+00:00"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.597Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -255,9 +255,9 @@ func ExampleGalleryApplicationVersionsClient_Get_getAGalleryApplicationVersion()
 	// 	Properties: &armcompute.GalleryApplicationVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryApplicationVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00+00:00"); return t}()),
+	// 			EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t}()),
 	// 			ExcludeFromLatest: to.Ptr(false),
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.5972568+00:00"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.597Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -332,9 +332,9 @@ func ExampleGalleryApplicationVersionsClient_NewListByGalleryApplicationPager() 
 		// 			Properties: &armcompute.GalleryApplicationVersionProperties{
 		// 				ProvisioningState: to.Ptr(armcompute.GalleryApplicationVersionPropertiesProvisioningStateSucceeded),
 		// 				PublishingProfile: &armcompute.GalleryApplicationVersionPublishingProfile{
-		// 					EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00+00:00"); return t}()),
+		// 					EndOfLifeDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-07-01T07:00:00.000Z"); return t}()),
 		// 					ExcludeFromLatest: to.Ptr(false),
-		// 					PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.5972568+00:00"); return t}()),
+		// 					PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2019-06-21T17:13:57.597Z"); return t}()),
 		// 					ReplicaCount: to.Ptr[int32](1),
 		// 					StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 		// 					TargetRegions: []*armcompute.TargetRegion{