+++ packages/autorest.gotest/test/integrationtest/output/compute/armcompute/galleryimageversions_client_example_test.go
@@ -95,7 +95,7 @@ func ExampleGalleryImageVersionsClient_BeginCreateOrUpdate_createOrUpdateASimple
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -234,7 +234,7 @@ func ExampleGalleryImageVersionsClient_BeginCreateOrUpdate_createOrUpdateASimple
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -376,7 +376,7 @@ func ExampleGalleryImageVersionsClient_BeginCreateOrUpdate_createOrUpdateASimple
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -510,7 +510,7 @@ func ExampleGalleryImageVersionsClient_BeginCreateOrUpdate_createOrUpdateASimple
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -644,7 +644,7 @@ func ExampleGalleryImageVersionsClient_BeginCreateOrUpdate_createOrUpdateASimple
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -739,7 +739,7 @@ func ExampleGalleryImageVersionsClient_BeginUpdate_updateASimpleGalleryImageVers
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -827,7 +827,7 @@ func ExampleGalleryImageVersionsClient_BeginUpdate_updateASimpleGalleryImageVers
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -895,7 +895,7 @@ func ExampleGalleryImageVersionsClient_Get_getAGalleryImageVersionWithReplicatio
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -983,7 +983,7 @@ func ExampleGalleryImageVersionsClient_Get_getAGalleryImageVersionWithSnapshotsA
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -1052,7 +1052,7 @@ func ExampleGalleryImageVersionsClient_Get_getAGalleryImageVersionWithVhdAsASour
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -1123,7 +1123,7 @@ func ExampleGalleryImageVersionsClient_Get_getAGalleryImageVersion() {
 	// 	Properties: &armcompute.GalleryImageVersionProperties{
 	// 		ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 	// 		PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+	// 			PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 	// 			ReplicaCount: to.Ptr[int32](1),
 	// 			StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 	// 			TargetRegions: []*armcompute.TargetRegion{
@@ -1222,7 +1222,7 @@ func ExampleGalleryImageVersionsClient_NewListByGalleryImagePager() {
 		// 			Properties: &armcompute.GalleryImageVersionProperties{
 		// 				ProvisioningState: to.Ptr(armcompute.GalleryImageVersionPropertiesProvisioningStateSucceeded),
 		// 				PublishingProfile: &armcompute.GalleryImageVersionPublishingProfile{
-		// 					PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00Z"); return t}()),
+		// 					PublishedDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2018-01-01T00:00:00.000Z"); return t}()),
 		// 					ReplicaCount: to.Ptr[int32](1),
 		// 					StorageAccountType: to.Ptr(armcompute.StorageAccountTypeStandardLRS),
 		// 					TargetRegions: []*armcompute.TargetRegion{